+++ database/seeds/TestingDatabaseSeeder.php
@@ -3,6 +3,7 @@
 use Faker\Factory as Faker;
 use Illuminate\Database\Seeder;
 use JobApis\JobsToMail\Models\User;
+use JobApis\JobsToMail\Models\Recruiter;
 use JobApis\JobsToMail\Models\Search;
 use JobApis\JobsToMail\Models\Token;
 
@@ -14,6 +15,7 @@ class TestingDatabaseSeeder extends Seeder
         $this->createActiveUsers();
         $this->createDeletedUsers();
         $this->createUnconfirmedUsers();
+        $this->createRecruiters();
     }
 
     private function createActiveUsers($num = 10)
@@ -42,6 +44,11 @@ class TestingDatabaseSeeder extends Seeder
                     factory(Token::class)->make()
                 );
             });
+    }
+
+    private function createRecruiters($num = 100)
+    {
+        return factory(Recruiter::class, $num)->create();
     }
 
     private function createUnconfirmedUsers($num = 10)