+++ changelog.md
@@ -14,6 +14,8 @@ All Notable changes to `jobs-to-mail` will be documented in this file.
   - Unsubscribe from all searches
   - View this user's active searches
 - New page to view a user's searches
+- Filtering out recruiter listings:
+  - `recruiters` database table, seeder, model
 
 ### Fixed
 - Plaintext email job listings were showing up as blank.